+++ Common/src/main/java/fuzs/sheepvariety/common/SheepVariety.java
@@ -1,13 +1,13 @@
-package fuzs.sheepvariety;
+package fuzs.sheepvariety.common;
 
 import fuzs.puzzleslib.common.api.core.v1.ModConstructor;
 import fuzs.puzzleslib.common.api.core.v1.context.DataPackRegistriesContext;
 import fuzs.puzzleslib.common.api.event.v1.entity.ServerEntityEvents;
 import fuzs.puzzleslib.common.api.event.v1.entity.living.BabyEntitySpawnCallback;
 import fuzs.puzzleslib.common.api.event.v1.entity.player.PlayerInteractEvents;
-import fuzs.sheepvariety.handler.SheepSpawnVariantHandler;
-import fuzs.sheepvariety.init.ModRegistry;
-import fuzs.sheepvariety.world.entity.animal.sheep.SheepVariant;
+import fuzs.sheepvariety.common.handler.SheepSpawnVariantHandler;
+import fuzs.sheepvariety.common.init.ModRegistry;
+import fuzs.sheepvariety.common.world.entity.animal.sheep.SheepVariant;
 import net.minecraft.resources.Identifier;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;