+++ Common/src/main/java/fuzs/sheepvariety/common/client/renderer/entity/layers/SheepVariantWoolUndercoatLayer.java
@@ -1,17 +1,17 @@
-package fuzs.sheepvariety.client.renderer.entity.layers;
+package fuzs.sheepvariety.common.client.renderer.entity.layers;
 
 import com.google.common.collect.Maps;
 import com.mojang.blaze3d.vertex.PoseStack;
-import fuzs.sheepvariety.client.model.geom.ModModelLayers;
-import fuzs.sheepvariety.client.renderer.entity.state.SheepVariantRenderState;
-import fuzs.sheepvariety.world.entity.animal.sheep.SheepVariant;
+import fuzs.sheepvariety.common.client.model.geom.ModModelLayers;
+import fuzs.sheepvariety.common.client.renderer.entity.state.SheepVariantRenderState;
+import fuzs.sheepvariety.common.world.entity.animal.sheep.SheepVariant;
 import net.minecraft.client.model.EntityModel;
 import net.minecraft.client.model.animal.sheep.SheepModel;
 import net.minecraft.client.model.geom.ModelLayers;
 import net.minecraft.client.renderer.SubmitNodeCollector;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.client.renderer.entity.RenderLayerParent;
-import net.minecraft.client.renderer.entity.layers.RenderLayer;
+import net.minecraft.client.renderer.entity.layers.SheepWoolUndercoatLayer;
 import net.minecraft.client.renderer.entity.state.SheepRenderState;
 import net.minecraft.world.item.DyeColor;
 
@@ -20,11 +20,11 @@ import java.util.Map;
 /**
  * @see net.minecraft.client.renderer.entity.layers.SheepWoolUndercoatLayer
  */
-public class SheepVariantWoolUndercoatLayer extends RenderLayer<SheepRenderState, SheepModel> {
+public class SheepVariantWoolUndercoatLayer extends SheepWoolUndercoatLayer {
     private final Map<SheepVariant.ModelType, SheepModel> models;
 
-    public SheepVariantWoolUndercoatLayer(RenderLayerParent<SheepRenderState, SheepModel> renderLayerParent, EntityRendererProvider.Context context) {
-        super(renderLayerParent);
+    public SheepVariantWoolUndercoatLayer(RenderLayerParent<SheepRenderState, SheepModel> renderer, EntityRendererProvider.Context context) {
+        super(renderer, context.getModelSet());
         this.models = bakeModels(context);
     }
 
@@ -39,18 +39,21 @@ public class SheepVariantWoolUndercoatLayer extends RenderLayer<SheepRenderState
 
     @Override
     public void submit(PoseStack poseStack, SubmitNodeCollector submitNodeCollector, int lightCoords, SheepRenderState state, float yRot, float xRot) {
-        if (!state.isInvisible && (state.isJebSheep || state.woolColor != DyeColor.WHITE) && !state.isBaby
-                && ((SheepVariantRenderState) state).variant != null) {
-            EntityModel<SheepRenderState> entityModel = this.models.get(((SheepVariantRenderState) state).variant.assetInfo()
-                    .model());
-            coloredCutoutModelCopyLayerRender(entityModel,
-                    ((SheepVariantRenderState) state).variant.assetInfo().undercoat().texturePath(),
-                    poseStack,
-                    submitNodeCollector,
-                    lightCoords,
-                    state,
-                    state.getWoolColor(),
-                    1);
+        if (((SheepVariantRenderState) state).variant != null) {
+            if (state.isBaby) {
+                super.submit(poseStack, submitNodeCollector, lightCoords, state, yRot, xRot);
+            } else if (!state.isInvisible && (state.isJebSheep || state.woolColor != DyeColor.WHITE)) {
+                EntityModel<SheepRenderState> entityModel = this.models.get(((SheepVariantRenderState) state).variant.assetInfo()
+                        .model());
+                coloredCutoutModelCopyLayerRender(entityModel,
+                        ((SheepVariantRenderState) state).variant.assetInfo().undercoat().texturePath(),
+                        poseStack,
+                        submitNodeCollector,
+                        lightCoords,
+                        state,
+                        state.getWoolColor(),
+                        1);
+            }
         }
     }
 }