+++ Common/src/main/java/fuzs/sheepvariety/common/data/ModDataPackRegistriesProvider.java
@@ -1,9 +1,9 @@
-package fuzs.sheepvariety.data;
+package fuzs.sheepvariety.common.data;
 
 import fuzs.puzzleslib.common.api.data.v2.AbstractDatapackRegistriesProvider;
 import fuzs.puzzleslib.common.api.data.v2.core.DataProviderContext;
-import fuzs.sheepvariety.init.ModRegistry;
-import fuzs.sheepvariety.world.entity.animal.sheep.SheepVariants;
+import fuzs.sheepvariety.common.init.ModRegistry;
+import fuzs.sheepvariety.common.world.entity.animal.sheep.SheepVariants;
 
 public class ModDataPackRegistriesProvider extends AbstractDatapackRegistriesProvider {